rename github to GitHub
