+++ src/routes/app/(flows)/claim-project/steps/add-ethereum-address/add-ethereum-address.svelte
@@ -35,7 +35,7 @@
   import assert from '$lib/utils/assert';
   import ethAddressItem from '$lib/components/drip-list-members-editor/item-templates/eth-address';
   import Checkbox from '$lib/components/checkbox/checkbox.svelte';
-  import GitHub from '$lib/utils/github/github';
+  import GitHub from '$lib/utils/github/GitHub';
   import { gql } from 'graphql-request';
   import { Octokit } from '@octokit/rest';
